Arrange device implementation
